Additions.
Great job octaviogutierrezmc3.  I Changed the format a little, but most of that is personal preference. The bulk of what I did is I added a few methods for Altitude, and Is crashed accord to the book on the next page.  I also changed the earths radius.  the book says its 12,756,300, I think its about half that.  Which would mean that its really the diameter.  I tried to test it but the satellite crashes on the first step of time.  So I have something wrong, I am still working on some kinks.  let me know if you see what the issue is.